refactor: 폴더구조를 MVC 기반으로 변경
