+++ src/main/java/racingcar/model/CarPower.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.model;
 
 import camp.nextstep.edu.missionutils.Randoms;
 
@@ -9,18 +9,14 @@ public class CarPower {
 
 	private final int power;
 
-	private CarPower(int power) {
+	//쉬운 테스팅을 위한 protected 생성
+	protected CarPower(int power) {
 		this.power = power;
 	}
 
 	public static CarPower createRandomPower() {
 		int randomPower = Randoms.pickNumberInRange(POWER_MIN, POWER_MAX);
 		return new CarPower(randomPower);
-	}
-
-	//쉬운 테스팅을 위한 protected 생성
-	protected static CarPower createPower(int power) {
-		return new CarPower(power);
 	}
 
 	public boolean canMove() {
