+++ src/test/java/racingcar/model/CarTest.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.model;
 
 import static org.assertj.core.api.AssertionsForClassTypes.*;
 import static org.junit.jupiter.api.Assertions.*;
@@ -8,17 +8,20 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.NullAndEmptySource;
 
+import racingcar.model.Car;
+import racingcar.model.CarPower;
+
 public class CarTest {
 
 	@DisplayName("자동차가 잘 움직이는지 테스트")
 	@Test
 	void move() {
 		Car move = new Car("move");
-		move.moveOrStop(CarPower.createPower(4));
+		move.moveOrStop(new CarPower(4));
 		assertEquals(new Car("move", 1), move);
 
 		Car noMove = new Car("noMov");
-		noMove.moveOrStop(CarPower.createPower(3));
+		noMove.moveOrStop(new CarPower(3));
 		assertEquals(new Car("noMov", 0), noMove);
 	}
