+++ src/test/java/racingcar/model/TryCountTest.java
@@ -1,9 +1,11 @@
-package racingcar;
+package racingcar.model;
 
 import static org.assertj.core.api.Assertions.*;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+
+import racingcar.model.TryCount;
 
 public class TryCountTest {
